requirements.txt
asgiref==3.3.1
Django==3.1.3
django-taggit==1.3.0
django-tags-input==5.0.0
Markdown==3.3.3
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
sqlparse==0.4.1